Tilemap Collision Behavior w/ Jump: Work In Progress
Trying to do this by disabling the Player's collision box when they start a jump such that the terrain can't cause a collision with the Player, and then I reenable it at the end of the jump. This works and doesn't cause a lagspike (fiddling with the Tileset's physics layers causes lag spikes in-game due to how large the file is).

Having a hard time with landing on the platform/ledge. I need to figure out a good way to keep track of the Player's altitude and manage the z-index appropriately. I'm thinking I'll have to go backwards with the z-index as the altitude increases so that assets that are closer/lower in elevation get drawn above assets that are further/higher.

I need to:
- Allow upward movement while jumping
- Measure altitude (z-index? or maybe a separate value that can impact z-index?)
- Fix z-indexing of Characters when they are on top of a ledge
- Map out z-index chart