Tile Manager & Tile Scripts Created
I organized my code for tiles so that they'd become much easier to work with. I was struggling with the jumping code, and I figure it'll be easier to work on it when the Tiles are easier to manager. This makes tile data and custom data a lot easier to fetch and i think i'll be able to better understand it now that i've got it formatted this way