Jumping Onto Elevated Surfaces: Work In Progress
I'm working on functionality to allow characters to jump onto ledges and elevated platforms. It's going okay, but I've put a lot of time into it and I'm not really seeing a super clear answer on how to make this work.

Currently there is an Area2D attached to the player's feet that will detect collisions between the feet and the tilemap (typically colliding with ledges). When the Player stands on a ledge tile, the code is supposed to set an 'on_platform' variable to true- but this isn't always working.

While that on-platform variable is true, the player should be able to walk around the platform and walk off of it to trigger a quick fall animation as they return to the ground.

It is almost working as it needs to but it needs a little bit more work.