Jumping Onto Elevated Surfaces: Still Work in Progress
This is about 80% complete now. most jumps seem to work fine, Its very buggy though, and I need to figure out a way to refine it otherwise I'll probably redo the whole thing.

Redoing it wouldn't be the end of the world because I've learned a fair amount about collisions and have gained more Godot knowledge, so I think it could be good to start from square one- but only if necessary.

The current issues with the jumping are primarily with the highest points of hills/cliffs. Every ledge tile currently has no collision barrier along their tops- only on their sides and bottom. The exception to this is ledges that are at the highest point because the player or character could walk off that tile since there is no ledge above to keep it in-bounds.

It is also just generally glitchy with the collision barrier bodies and feet. I think I'm going to watch a video for top-down jump physics to see if I could replicate, personalize, and expand upon a guide.