Memory Leak Fix & New Jumping/Collisions Plan
I discovered a memory leak *dun dun dun*!! It was caused by the absurd number of tiles I was creating after I made a class to temporarily hold data for them as objects. The objects being created were not removed after use, and they quickly accumulated to the hundreds of thousands.
I began debugging, and I learned a lot. I became more familiar with Godot's debugger menu, though I still don't quite understand the ins and outs. I also, most importantly, learned how to properly unload an asset from memory instantly and effectively. It's all about ensuring that no references exist after you call ".free( )" on the object, and I set the object to "null" afterwards to ensure the value is cleared out. It got my objects down from 300k+ to 2600ish. Now that I know how to do this, I hope to go back through old code and free objects wherever possible.

-

The current jumping system is mid and meager. It is made with a really sloppy jump system that depends on a narrow collision space. I've learned a lot now about collision layers and masks, and I realize that I can add multiple physics layers for multiple states of tiles with respect to characters. For example, a tile could have all of its physics layers active, but each of them corresponds to one of the characters states. If the character is above the tile, they'll only enable physics layer 1 and mask 1, which will correspond to an appropriate collision physics layer for that state on the tilemap. I also have plans to utilize a backwards movement onto the z-axis as a means of tracking altitude for jumping and falling from ledges and platforms. Gonna work on it and see what happens. Here's a good checkpoint to leave off at.