All Nodes Updated & Corresponding Scripts Load Nodes Effectively
I added the collision body and feet to every possible kind of character I have so far, and I learned how to use the super() constructor to make it much simpler to instantiate the nodes in subclasses of Character. I'm seeing the benefits of my code structure work in real time!