New Jump System: Getting There...
This one is slowly coming along.

It has pretty good collision detection. I figured out how to switch the active layers and mask through code, so right as a character jumps their colliders switch layers/masks. When the character starts falling after reaching the peak of their jump, the Area2D on the feet activates. It is designed to be two pixels lower than the actual foot, so it detects a platform early, switches the character's collision layer/mask back to what it was before, and allows the feet collider to land on the tile it detected. This works pretty smoothly actually, and I haven't been able to replicate any errors with the feet anymore- but I did replicate one with the body collider. I'll need to find a solution for that.

I have an altitude value that I'm adjusting based on jump rise/fall height. It works okay, and I'm able to simulate fairly accurate falls from elevations, but it also is slightly imprecise due to the conversion for altitude from normal y-coords.
I think I'm gonna switch it so that whenever you jump, your z-index decreases by 1 (to move you higher) and if you land on something it stays decreased. That will be an altitude change of 1.
If you jump again your z-index would decrease again, but then if you land at the same height you initially jumped from your z-index would revert the latest decrease.
If you jump up but then fall back down to the ground from any altitude, your z-index will become 1.
If you walk off of a ledge, your z-index will increase once for that first tile, and then it will increase again for every tile (16 px) you fall after that. The body_exited( ) area2D callback could be useful for triggering this-.
Falling would be a lot more refined with this z-axis system, because then whenever falling starts, it will stop after the character has fallen exactly 16 * -z-index pixels.

This will require a function to manage the Character's state while on the platform. It might be doable just through the ledge tile callbacks, but I'm not sure. I'm also not sure if having ledge_grass set will be useful, but I should definitely try to find a way to not have to do that.

If I can get the character able to jump/fall and land at the exact proper heights they should based on character a fake z-axis with the z-index, and if I can improve collisions for a near-guaranteed success rate, I'll consider this done.