New Jump System: HUGE Improvement, Still Working Though
I implemented a shadow that follows an adjustable landing height variable. The Player can also now move in all four directions when jumping. The shadow helps show where the Player's jump will land. If I can figure out why the Camera2D stutters as the character lands rather than remaining smooth, then I'd like to code it such that the camera will switch focus from the player to the shadow when they jump, and when the jump ends I'd like it to switch back. It is easier to control and steer the jump when the camera is focused on the shadow, but I want the Player to be centrally focused at all other points.

I need to improve collisions a bit more. Sometimes the Player's colliders can get stuck inside of the tile colliders. I think I could improve this with better structured physics layers and a well thought out design. I can also try to smooth out collisions in code potentially by resolving a collision before it gets stuck.

The only other thing I need to implement is the z-axis/altitude system, which will also include logic for falling. I think I'll be able to depend on the jump_landing_height variable I made in character.gd for a lot of these z-axis calculations.