New Jump System: Z-Index Altitude Foundation In-Place
I have laid the foundation down for the altitude functionality to exist through the z-index of each Character. It is still a bit finnicky right now and needs tweaking, but it did work in some test scenarios. It depends on recursive function calls to find a ledge_front tile, then to find how many ledge_wall tiles are beneath that ledge_front tile, and that number will equal the altitude of the tile. That information allows for easy adjustments to the Characters' z-indexes.

Remaining tasks for full functionality include:
- Make necessary tweaks to altitude calculation system.
- Add proper collisions and z-index perspective logic.
- Fix camera jitter while focused on the shadow of the character
- Implement falling functionality based on altitude (jumping off the sides should just set altitude to 0 when they land on ground, but jumping off the front/back should position the Character according to wall height)