New Jump System: Z-Index Altitude System Refined, Needs Improvements Down the Line
I have made the z-index altitude system functional. It will detect a Character's altitude based on the heights of the start and end tiles of their jumps.

Down the line, I plan to work off of this system to implement falling mechanics as well as logic to ensure that a Character can't jump up onto a ledge that's too high from any side (front, rear, or sides). As of now this is enough of a baseline functionality for testing.

Remaining tasks for full functionality include:
- Add proper collisions and z-index perspective logic.
- Fix camera jitter while focused on the shadow of the character
- Implement falling functionality based on altitude (jumping off the sides should just set altitude to 0 when they land on ground, but jumping off the front/back should position the Character according to wall height)