New Jump System: Y-Sorting In Progress
After much trial and much error, I decided to watch a YouTube video to figure out a good way to handle z-indexing the characters and assets automatically in a manner that looks smooth.

I found a video discussing 'y-sorting' which is the exact same thing I've been describing as z-indexing / perspective. I enabled it on the Player and the tilemap, and it seems to be working alright right now. I think it will need to be developed further alongside collision logic, which I currently don't have the brain power to figure out.

I will try to develop those two simultaneously tomorrow and hopefully have a finished product.