New Jump System: FOUNDATIONALLY COMPLETE!!!
"Jump System" isn't even all that is complete now- it's actually more like a "Terrain Navigation System".

Just over two weeks ago, I began an experiment to try and code functionality for jumping onto the terrain cliff/ledge tiles that Ryan had drawn. This was mainly intended to be a quick "get your feet wet" kind of experiment, but I actually ended up making pretty good progress and coming up with lots of ideas so I decided to stick with it.

Although I was seeing results and making good progress in my experiments with the player character, none of the progress was actually tangible because it couldn't be replicated for the entire game. The ways that I was manipulating the TileMapLayers and the collision logic wasn't scalable.

Despite not making scalable changes, I learned a ton about so many different topics and concepts. I decided to power through and continue working towards fulfilling this "jump system" experiment in a scalable manner while all of my newfound knowledge was still fresh in my head.

Over the course of the past two weeks, I learned and did all of the following:

- Learned about TileMapLayers and TileSets. This included deepening my understanding of terrain sets, custom data layers, physics layers, and y-sort.
After learning and after some trial and error, the implementation here has two different physics layers on the mega_tilemap.tres TileSet. One of those layers is for when a character is on the ground and the other is for while they're on a platform.
The code dynamically updates the Character's collision value and which physics colliders (body and feet) are enabled at any given moment based on the tiles it is standing on or nearest to.
The code will also switch certain tiles to alternative tiles such that their y-sort origin works effectively based on the Character's on-platform status.

- Learned more about physics and worked more on jumping logic. I made it so that the Character can move in any direction while they're jumping now, and there's a drop shadow that allows the Character to better see where they're landing.

- Learned about memory leaks. I had created a Tile class so that it'd be easier to access specific values that I needed from Tiles. I was creating them rapidly (because a whole grid of tiles surrounding the Character need to be processed at once), but they were persisting in memory as orphan objects and weren't being removed faster than they were being created. I ended up learning how to unload objects properly which will be incredibly useful down the road.

There was probably even more that I did but didn't account for. This was a crazy sprint- definitely the hardest stuff I've done so far for this game, but I think it will be very valuable because now we will be able to design the map with more detail than just a floor plan. We can build up any terrain with this environment interaction logic, and I think it will also make it much easier for adding buildings and other assets like trees and rocks as well.

I plan to spend time tomorrow going through the tons of code I've made during the past two weeks and tidying it up a bit. More comments, revise old inaccurate comments, and remove any code that is no longer useful.