NatureAsset Interaction Code
I added code that will determine if the Player is close to a Tile on the environment layer (trees, rocks, bushes, etc.). If the Player is close, a NatureAsset is created and instantiated at the exact coordinates of the Tile and is added to the interactables list to be processed like other interactables.

Another algorithm will determine if the closest interactable is a NatureAsset, and if so, the algorithm calls the closest interactable's interaction function (which is currently set to print "berries collected!" for a test.

I can't commit this until I ensure that the NatureAsset script is adequately prepared. It will be extending Interactable, which has an associated Scene with Nodes, but that is designed for Characters. I need to restructure my Scenes now that I've done work with Tiles and know how assets will be loaded in primarily.