Altitude and Collision Fixes/Updates
The altitude and collision code was in a weird state after an issue I had where I didn't push a commit from my Macbook, then I made changes in the same branch on my desktop, so when I went on my Macbook the changes were gone.

I updated the code so that the TileManager can calculate a Character's altitude based on the Tile that they are currently standing on.

I updated game.gd to have a signal processing when a Character wants to update their altitude, and also patched a memory leak by unloading a tile that wasn't being unloaded previously.

I also modified and added the collision layers for each of the terrains that Ryan drew (grass/dirt, cake/frosting, dirt/soil). I still need to add the collision layers for environments on layer 3, and I also need to add collision layers for fences on both 1 and 3.