Fixed Ryan's Cliff/Ledge Terrains for Garden & Sweets World
We both worked on the mega_tilemap.tres file the other day and we had to do a merge of our work. During the merge, I think I removed some of Ryan's changes and kept mine, and when I reopened the file after the merge Ryan's terrains weren't as they were before.

I had to redraw the terrain grid shape onto the garden and sweets world cliffs, but it is working okay now and will be ready for Ryan to work with later.