EnvironmentAsset Progress: Creating Associated Object for Nature Tiles
I came up with the idea to add a "connected_object" custom data field to each tile. This allows for each tile to connect to an actual object that can be interacted with in more ways than just physics collisions. For example, a bush tile that can be collided with can't be interacted with, but if it has an associated object at the exact same location as the tile, the object can be interacted with.

The nature asset tile callback function now checks if a nature asset tile already has an associated EnvironmentAsset object, and if it doesn't, one will be created.

I need to add more functionality to the EnvironmentAsset object. I may need to create subclasses of EnvironmentAsset for Tree, Bush, and Rock, but if I can cover the functionality of all three with just EnvironmentAsset then I won't. I will, however, likely need to create a separate .tscn file for Tree, Rock, and Bush so that their 'Press 'E' to Interact' labels can appear at the correct height.

environment-asset.gd:
- Switched to extend InteractableAsset rather than InteractableCharacter

- Created the init function, which is currently only setting global position of the EnvironmentAsset to whatever the given coordinates are (typically given from the corresponding nature tile).

game-tile.gd:
- Added a 'connected_object' variable to store an associated object to a tile.

- Updated the init function to properly initialize the new 'connected_object' variable.

interactable-asset.gd:
- Updated the class name and comments

game.gd:
- Removed previous experimental EnvironmentAsset code.

- Added a character parameter for the food buddy killing target signal callback function.

tile-manager.gd:
- Updated the nature asset tile callback function to check if a tile has a connected object, and if not, one is created and assigned.