EnvironmentAsset (Bushes, really): COMPLETE
I finally finished implementing the logic to allow bushes to be interacted with. I had been working to allow EnvironmentAssets (Trees, Rocks, and Bushes) to be interacted with, but after discussing with Ryan, we decided it'd only be necessary for bushes to be interacted with for the capstone deadline.

When the player or a character walks close enough to a bush tile, code is triggered to create an invisible object right on top of the tile. The object represents the bush, and it will store all of the logic/behaviors of the bush- including the amount of berries it currently has, transferring berries to the player/character that interacts with it, and regenerating berries.

mega_tilemap.tres:
- Switched 'nature_asset' to 'environment_asset', and then switched bushes to have 'bush' as their tile type.

bush.tscn:
- Created a prefab/Scene for the bush that has an interaction hitbox, damage hitbox, press 'e' to interact label, and I ensured that it would fit and match the position and size of the bush despite not having an actual skin (because the tile serves as the skin for the object when its created).

malick.tscn and sally.tscn:
- Disabled the 'Press 'E' to Interact' label by default.

bush.gd:
- Created this file to extend InteractableAsset. It will define all of the behavior and logic for a bush.

- The 'interact_with_player' function is defined to simply print "berry retrieved", as this was testing that the pipeline is working after the bush is created. I will update this to have the berry collection logic for the player/character that interacts with it.

environment-asset.gd:
- Removed this because I am only defining behavior for bushes, so a file for all environment assets isn't necessary.

interactable-asset.gd:
- Added variables for 'in_range', 'center_point', and 'paused' because these are used when processing Interactables.

- center_point isn't actually the center point right now- I need to determine the height of the bush and decrement the y-value by half the height in order to get the real center coord. It is functioning without this being the actual location though.

player.gd:
- Turned off bottom point printing for debugging