Bug Fixes (y-sort/z-index & jumping)
I wanted to fix the slight bugs that existed when jumping. There were problems with y-sorting against the EnvironmentAssets, so I came up with a solution to increment the player's z-index if the starting point of the jump was in front of one of those assets. This might be a scalable solution for terrain tiles, too. Now when the player jumps while in front of a tree, for example, they will appear in front of the tree for the entire duration of the jump rather than clip behind it in mid-air briefly.

The shadow underneath the player is fixed now. It wasn't really following the player's movements very well, and the code for it was quite messy. It also y-sorts now.

mega-tilemap.tres:
- I made the collision barriers for trees and rocks a bit higher so that when the player jumps while behind them, they and their shadow don't clip through the asset briefly while landing.

player.tscn:
- Removed the old shadow, added a new one, and enabled y-sort on it.

game-character.gd:
- Updated the shadow code with the newly added shadow to player.tscn. It no longer needs to be offset by 22 px.

player.gd:
- Implemented more shadow code so that the position of the shadow is adjusted by vertical movement during the jump.

- Added one more shadow line to ensure it returns to exactly the player's feet at the end of the jump. I had tried to put that line of code at the end of the 'jump_end' function in GameCharacter.gd, but it wouldn't always reset the shadow to the appropriate location when the code was there.

tile-manager.gd:
- Incremented the character's z-index by 1 if the start of their jump was in front of an EnvironmentAsset so that they wouldn't clip through the asset in mid-air due to y-sorting.