Bug Fixes (In Progress): Jumping & Tile Processing
I'm trying to fix two important bugs. One allows the player to jump through EnvironmentAssets. I made a little bit of progress towards this one. The character is not able to jump through the asset if they're standing in front of it, but they sometimes can glitch through it with a running start. They aren't able to jump through them from behind (maybe with a running start they can), but it is buggy differently because they'll keep falling in-place if they try to move downward to go through the asset. Then, once they step to the side, they fall down to wherever their shadow moved to (which can be infinitely far). I couldn't find a solution for that.

The other bug in is that tiles are only being processed in the town center area. The tile callback functions don't seem to be executing for the other areas aside from the ground callback. I'm trying to implement functionality to switch the active tilemaps for the ground, terrain, and environment to match the location the player is in. It is taking longer than I anticipated. The TileManager system might need to be updated, as it seems to be using quite a lot of computational power (particularly the process_nearby_tiles function).

mega-tilemap.tres:
- Duplicated the EnvironmentAsset's physics layer 1 into physics layer 2 such that the player will collide with them when jumping.

- Added a new custom tile data for 'location' (center, sweets, or garden) and painted each tile with their corresponding location.

game.tscn:
- Moved the player into the sweets world for testing.

world-map.tscn:
- Built a terrain in the sweets world for testing.

game-character.gd:
- Updated the collision value to be set to 2 instead of 6 such that the character will collide with EnvironmentAssets when jumping.

player.gd:
- Updated one of the debug statements such that it accurately prints the coordinates of the player's bottom point (which is oriented in its .tscn file to be right where their feet are).

game.gd:
- Added a dictionary to store every tilemap for each area.

- Updated the initialization of the GameTileManager's tilemaps to use the new dictionary.

- Stopped processing tiles for Malick and Sally (I noticed a significant framerate increase when I did this).

tile-manager.gd:
- Added a variable to store all of the tilemaps

- Updated a comment