TileManager Updates & Z-Index / Y-Sorting Improvements
My main focus today was getting the TileManager to work a bit more efficiently, as I realized that the amount of tiles I was processing before was excessive and actually causing the framerate of the game to drop. I also needed it to be in a good position such that I can translate the functionality of tiles, y-sorting, and collisions over to the other worlds aside from the town center.

Now the TileManager is working a lot better, and the Tile callback functions are both updating the character's z-index and loading in alternative tiles with different y-sorting. It may still need a few tweaks, but it is in a good enough spot and much better than before. Y-sorting is working for the Food Buddies now, but oddly it only started working once I disabled y-sorting for them. I'm worried because I have no idea why that works, so if it breaks I'll have no idea how to fix it, but that's a problem for another day it seems.

mega-tilemap.tres:
- Updated the town center's "ledge_platform" tiles to have a y-sort-origin of 8 so that the characters never clip through it.

- Updated the medium sized rock in the town center to have consistent physics layers.

malick.tscn and sally.tscn:
- Set the z-index of the AnimatedSprite2Ds to 0, it was 1 before and was causing issues with y-sorting in-game.

player.tscn:
- Moved the shadow higher so that it wouldn't clip through ground tiles with y-sorting (and apparently accidentally added a point).

game.tscn:
- Moved Malick and Player around while testing.

food-buddy-fusion, food-buddy, food-citizen, and food-enemy.gd:
- Changed 'update_location_points( )' to 'update_dimensions' after removing the center_point variable.

game-character.gd:
- Removed the center_point variable, as it seems to be better to track a character's global position (feet tracking feet).

- Changed the 'update_location_points' function to 'update_dimensions' now that it only updates width and height since center_point was removed.

interactable-asset.gd:
- Removed center_point code and replaced it with a pass statement.

player.gd:
- Changed 'update_location_points( )' to 'update_dimensions' after removing the center_point variable.

game.gd:
- Enabled processing for Malick and Sally again after the TileManager.gd changes.

- Set the Player and Sally to only process two tiles higher than the tile they're currently standing on because they aren't tall enough to need three tiles above them processed.

- Changed 'center_point' to 'global_position' in various areas

tile-manager.gd:
- Revamped the 'process_nearby_tiles' function. It was inefficient and processing redundant tiles (and most likely not even processing the tiles I thought it was).

- That function now processes rows of three tiles. One row on the y-level where the character is standing, one row below that, one row above that, and optionally it can process any number of rows higher starting at two tiles above where the character is standing.

- Added some notes in comments to that function for more clarity (and my future sanity).

- Updated several tile callback functions to update the character's z-index in certain situations (some tiles do it when the character is in front of them, others do it when the character is behind them).

- Some tiles don't update the z-index in their callbacks and only load in an alternative tile to support y-sorting with respect to the character. Some tiles do both.

- These callback functions may need further tweaks and refinements down the line.