Collision Updates & More Z-Index/Y-Sort Fine Tuning
I made it such that the player, Malick, and Sally each have their own three collision layers. 1, 2, 3 for player, 4, 5, 6 for Malick, and 7, 8, 9 for Sally. This is so they don't collide with one another, but can still collide with the environment the same way. I added new physics layers with the exact same collision shapes as the ones that already existed to make this change feasible.

I noticed some glitchiness with the z-indexing/y-sorting, although it was very far and few between now and much better than it was just two days ago. There is a problem with the ledge_back tiles because they temporarily set the z-index to -1, and I added a line that sets it back to 0 in that same function, so as long as the character encounters another ledge or environment asset after being near a ledge_back, they should be good. This may need a better/more concrete solution, but its fine for now.

mega-tilemap.tres:
- I duplicated physics layers and renumbered them to the new physics layers that I added such that the collision boxes would be identical.

bush.tscn and food-citizen.tscn:
- Set the z-index of the "press 'e' to interact label' back to 1- mistakenly reset this last time.

malick.tscn and sally.tscn:
- Set the z-index of the "press 'e' to interact label' back to 1- mistakenly reset this last time.

- Updated the body colliders dimensions to ensure that the characters can't move into positions that would screw up their y-sorting.

- Disabled both colliders by default (they are dynamically adjusted by the tile callback functions).

world-map.tscn:
- Removed a rock from behind a ledge_back tile, because when that tile sets the character's z-index to -1, it ends up making their heads go underneath any assets above them, so there needs to be some clearance.

malick.gd, player.gd, and sally.gd:
- Added an enum to store the characters' respective collision values for ground, midair, and platform.

tile-manager.gd:
- Added an extra position check for the ledge tiles so that they only get processed if the character is in line with that tile horizontally and if they are above the tile (necessary for ledge_back).

- Replaced any hardcoded numbers in the "set_collision_value" function calls to use the newly created enums (there may be other files of code where I'll have to do this too).

- Set the z-index to 0 at the end of the 'tile_callback_ledge' function, and the only time it doesn't run is when a 'ledge_back' tile is being processed (in which case the z-index is set to -1 so that the characters appear behind the grass and ledge).