Tile Logic Experimentation
I'm currently experimenting with logic that will update the tilemaps being processed dynamically, such that when the character is standing on a border of worlds, tiles from both world can be processed appropriately.

The issue is currently with my code. If a bush tile exists at map coords (-2, 19) on the environment tilemap in world center and a bush tile exists at map coords (-2, 19) on the environment tilemap in sweets world, the code cannot differentiate properly. Very lucky find on this bug.