Tile Processing in ALL Worlds & Coliseum Tile Data Set & Stupid Dialogue Issue
I spent the past few commits in a whirlwind trying to implement several things regarding the tile system at once. I have it functioning such that buildings can be entered and exited properly now, with the coliseum as a test subject. All of the tiles and tilemaplayers for the coliseum are set up with proper z-indexing and collisions, and they have updated tileset resources to limit the memory usage of the game.

The processing of tiles now occurs effectively in all worlds after sorting out a huge issue with coordinate spaces. Each tilemaplayer shared the same (0, 0) coordinate, so not every coordinate pair was unique and it was leading to odd things such as a bush from sweets world at (0, 0) spawning at the (0, 0) of world center even though in the world center no bush was at that point.

Dialogue was fixed for the eighth time. Maybe that will seal the deal.

coliseum-exterior.tres & coliseum-interior.tres:
- Added the last necessary z-index and collision data.

All other .tres files in this commit:
- Updated to use a script called "dialogue_resource.gd" such that it won't override Godot's internal "dialogue.gd" file any longer. The scripts are identical other than the name.

game.tscn:
- Moved Player, Malick, and Sally around while testing.

world-map.tscn:
- Updated lots of stuff. I don't remember and the diff is too large.

player.gd:
- Enabled debug printing while testing.

game.gd:
-  Enabled processing for Malick and Sally

Dialogue.gd:
- Removed for the eighth time.

dialogue-resource.gd:
- Same as dialogue.gd, just different name

tile-manager.gd:
- Removed print debug statements