Building Dynamicability: COMPLETE (Also file organization)
I have successfully created a process for dynamically loading building objects into the game at the location of a building tilemaplayer. I added two black tiles to each buildings' color palette. One's tile_type became 'door', the other 'building'. I added a custom data field to the palette resource .tres files for 'global_position', and I manually set that value for the coliseum's 'building' tile and for candy castle's 'building' tile.
I drew with the door tile in almost every tile that makes up the door of each building, but I added a few 'building' tiles to each door such that they can blend in with the rest of the doors while also being able to trigger the spawning of the object.
tile-manager.gd has a callback for tile types of 'building', so when the player is close enough to the door and processes the 'building' tile, the callback is triggered, a building object is generated, a check is made to ensure the building hasn't already been instantiated, then the building is instantiated at the global_position deemed by the 'building' tile in the tilemaplayer.

I also organized some file names and folders.

.tres files:
- Added 'building' and 'door tile functionality ('building' tile triggers the instantiation of a building object at the correct location.

- Updated collision, z-index, y-sort for new tiles such that transitions and walking around buildings function correctly.

building.tscn:
- Increased the interaction hitbox size.

building.gd:
- Removed the manual setting of global position.

game.gd:
- Added a list specifically for interactable-assets (these will mostly, if not all, be tile objects).

- Ensured the list is updated every frame.

- Only made one call to "get_tree()" and stored the result in a 'scene_tree' variable, rather than calling it every frame.

- Updated the '_on_tile_object_enter_game' callback such that the tile object's global position is set to the 'global_position' custom data if the tile is a building, otherwise it is just set to the global coordinates of the individual tile itself.
Also moved the iteration check to ensure no duplicates are created to the bottom of the function. This will change into a dictionary check eventually.

tile-manager.gd:
- I know that future me looking back on these commits must be terrified everytime they see changes to this file. This one is a monster and it takes so long to wrap your head around it.

- Fear not this time! It was just making the debugger stop yelling at me by prefixing an unused param with an underscore.