Path-Finding: Progress
I worked on path-finding and tested it with Malick and the Player. If there is an obstacle that Malick is above and the Player is beneath, Malick now knows how to maneuver around it. I'm hoping to refine this code a bit more and then I'll copy and modify it for the other cases (like if Malick were below and the Player were above, and also left/right).

game-character.gd:
- Added the current_path list and modified the target to be a GameCharacter, not just a Node2D.

malick.gd:
- Increased speed by 10

game.gd:
- Updated the move_towards_target function to work with paths.

- Probably can be tidied up a bit, but it is fairly well commented and is functioning okay.

tile-manager.gd:
- Added the generate_path function.

- Added a function called 'temp' (will name it later), which is called from the environment assets callback function. The 'temp' function adjusts the character's position when they're near objects in a way that will prevent them from colliding with the object.