Building Fixing
Fixed some issues with building resources (after we had adjusted the palettes for each tilemaplayer we lost tile data, so I replaced it). Currently there is a bug with the Candy Castle entering in which the surrounding tilemaps aren't turning off as they're supposed to. I need to have a better system of updating the 'active' tilemaps, and potentially have a Character hold references to the tilemaps they're currently active in. Mucho easier.

.tres files:
Added tile data to palettes