New Building Tilesheet Set
I set up the building tilesheet for the "Building Interiors" and "Building Exteriors" tilemaplayers for each of the three worlds. I had to redraw each building, interior and exterior- that sucked. I also added all of the tile data to the newly added corn tiles. They seem to be working fine