Godot Path-Finding Experimentation
I switched back to using Godot's path-finding because I was quickly humbled by my attempt to manually code the path-finding. I found better tutorial videos and spent a longer amount of time trying to get it to work. It's somewhat functional right now but there are performance issues and it is only working in one world at at time.