Path-Finding in One World - No Debugger Errors
I modified the code and tilemaps such that the debugger wouldn't have any problems with me. I also discovered the limitations of the navigation system, knowing now that I can really only have one world's worth of tiles loaded for the characters to navigate on without lagging the game too badly.

I might be in better shape if I attempt to do a few different things. If I try to switch the TileManager system such that active tilemaps are stored by the character rather than the system, that'd be much better and would eliminate some other bugs. Then I could see if condensing the mega tilesheet into multiple tilemaps would be beneficial for performance. Big pain in the butt, BUT, it would be the last of the large-scale functionality I'd have to worry about- so after this the tile manager should be complete.