Performance Experiment
I needed to determine if the path-finding was causing lag spikes or if it was due to having too much data in one tilemap. I created a whole new resource and tilesheet separate for the world center (this type of organization feels a lot better), but using it didn't fix the lag spikes. It also broke some of the z-indexing and y-sorting, which is very no bueno, so I'm going to stick to the OG tilemap.

On the bright side, tilemaps are now stored locally by the characters and characters can be in different worlds at the same time now. Thank the lords!